# packages/sdk/openapi.json
{
  "components": {
    "schemas": {
      "APIError": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/APIError.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorPayload"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "ErrorPayload": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "ProfessionalAuthCreateSessionRequest": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/ProfessionalAuthCreateSessionRequest.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "professionalId": {
            "type": "string"
          }
        },
        "required": [
          "professionalId",
          "phone",
          "password"
        ],
        "type": "object"
      },
      "ProfessionalAuthPasswordRecoveryData": {
        "additionalProperties": false,
        "properties": {
          "accepted": {
            "type": "boolean"
          },
          "requestedAt": {
            "type": "string"
          }
        },
        "required": [
          "accepted",
          "requestedAt"
        ],
        "type": "object"
      },
      "ProfessionalAuthPasswordRecoveryResponseBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/ProfessionalAuthPasswordRecoveryResponseBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/ProfessionalAuthPasswordRecoveryData"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
      "ProfessionalAuthRegisterRequest": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/ProfessionalAuthRegisterRequest.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "credentialNumber": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "professionalId": {
            "type": "string"
          }
        },
        "required": [
          "credentialNumber",
          "displayName",
          "password",
          "phone",
          "professionalId"
        ],
        "type": "object"
      },
      "ProfessionalAuthRequestPasswordRecoveryRequest": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/ProfessionalAuthRequestPasswordRecoveryRequest.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "professionalId": {
            "type": "string"
          }
        },
        "required": [
          "phone",
          "professionalId"
        ],
        "type": "object"
      },
      "ProfessionalAuthSessionData": {
        "additionalProperties": false,
        "properties": {
          "city": {
            "type": "string"
          },
          "credentialNumber": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string"
          },
          "isAuthenticated": {
            "type": "boolean"
          },
          "lastLoginAt": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "professionalId": {
            "type": "string"
          },
          "registeredAt": {
            "type": "string"
          }
        },
        "required": [
          "isAuthenticated",
          "phone",
          "professionalId"
        ],
        "type": "object"
      },
      "ProfessionalAuthSessionResponseBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/ProfessionalAuthSessionResponseBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/ProfessionalAuthSessionData"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
      "ProfessionalNotificationStateData": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/ProfessionalNotificationStateData.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "readIdsByProfessional": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": [
                "array",
                "null"
              ]
            },
            "type": "object"
          }
        },
        "required": [
          "readIdsByProfessional"
        ],
        "type": "object"
      },
      "ProfessionalNotificationsResponseBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/ProfessionalNotificationsResponseBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/ProfessionalNotificationStateData"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
      "ProfessionalPortalSessionData": {
        "additionalProperties": false,
        "properties": {
          "availableProfessionalIds": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "hasSnapshot": {
            "type": "boolean"
          },
          "lastActiveProfessionalId": {
            "type": "string"
          },
          "professionalId": {
            "type": "string"
          },
          "savedAt": {
            "type": "string"
          },
          "snapshot": {
            "additionalProperties": {},
            "type": "object"
          }
        },
        "required": [
          "hasSnapshot"
        ],
        "type": "object"
      },
...
      "SessionResponseBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/SessionResponseBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/ProfessionalPortalSessionData"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
      "UpsertProfessionalPortalSessionRequest": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/UpsertProfessionalPortalSessionRequest.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "professionalId": {
            "type": "string"
          },
          "snapshot": {
            "additionalProperties": {},
            "type": "object"
          }
        },
        "required": [
          "snapshot"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AdminSessionAuth": {
        "description": "Admin session cookie issued by POST /api/v1/admin/auth/session.",
        "in": "cookie",
        "name": "bidanapp_admin_session",
        "type": "apiKey"
      },
      "CustomerSessionAuth": {
        "description": "Customer session cookie issued by POST /api/v1/customers/auth/session.",
        "in": "cookie",
        "name": "bidanapp_customer_session",
        "type": "apiKey"
      },
      "ProfessionalSessionAuth": {
        "description": "Professional session cookie issued by POST /api/v1/professionals/auth/session.",
        "in": "cookie",
        "name": "bidanapp_professional_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Backend-generated OpenAPI contract for BidanApp.",
    "title": "bidanapp-backend",
    "version": "dev"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/notifications/professional": {
      "get": {
        "operationId": "get-professional-notifications-state",
        "parameters": [
          {
            "description": "Optional professional id to scope professional notification read state",
            "explode": false,
            "in": "query",
            "name": "professional_id",
            "schema": {
              "description": "Optional professional id to scope professional notification read state",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfessionalNotificationsResponseBody"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "ProfessionalSessionAuth": []
          }
        ],
        "summary": "Get persisted professional notification read state",
        "tags": [
          "Notifications"
        ]
      },
      "put": {
        "operationId": "upsert-professional-notifications-state",
        "parameters": [
          {
            "description": "Optional professional id to scope professional notification read state",
            "explode": false,
            "in": "query",
            "name": "professional_id",
            "schema": {
              "description": "Optional professional id to scope professional notification read state",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfessionalNotificationStateData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfessionalNotificationsResponseBody"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "ProfessionalSessionAuth": []
          }
        ],
        "summary": "Persist professional notification read state",
        "tags": [
          "Notifications"
        ]
      }
    },
...
    "/professionals/auth/password-recovery": {
      "post": {
        "operationId": "request-professional-password-recovery",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfessionalAuthRequestPasswordRecoveryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfessionalAuthPasswordRecoveryResponseBody"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Request a professional password recovery link",
        "tags": [
          "Professional Auth"
        ]
      }
    },
    "/professionals/auth/register": {
      "post": {
        "operationId": "register-professional-auth-account",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfessionalAuthRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfessionalAuthSessionResponseBody"
                }
              }
            },
            "description": "OK",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Bad Request"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Register a professional portal account",
        "tags": [
          "Professional Auth"
        ]
      }
    },
    "/professionals/auth/session": {
      "delete": {
        "operationId": "delete-professional-auth-session",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfessionalAuthSessionResponseBody"
                }
              }
            },
            "description": "OK",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "ProfessionalSessionAuth": []
          }
        ],
        "summary": "Revoke the authenticated professional session",
        "tags": [
          "Professional Auth"
        ]
      },
      "get": {
        "operationId": "get-professional-auth-session",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfessionalAuthSessionResponseBody"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "ProfessionalSessionAuth": []
          }
        ],
        "summary": "Get the authenticated professional session",
        "tags": [
          "Professional Auth"
        ]
      },
      "post": {
        "operationId": "create-professional-auth-session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfessionalAuthCreateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfessionalAuthSessionResponseBody"
                }
              }
            },
            "description": "OK",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Create an authenticated professional session",
        "tags": [
          "Professional Auth"
        ]
      }
    },
...
    "/professionals/portal/session": {
      "get": {
        "operationId": "get-professional-portal-session",
        "parameters": [
          {
            "description": "Optional professional id to load a specific portal snapshot",
            "explode": false,
            "in": "query",
            "name": "professional_id",
            "schema": {
              "description": "Optional professional id to load a specific portal snapshot",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseBody"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "ProfessionalSessionAuth": []
          }
        ],
        "summary": "Get persisted professional portal session",
        "tags": [
          "Professional Portal"
        ]
      },
      "put": {
        "operationId": "upsert-professional-portal-session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertProfessionalPortalSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseBody"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "ProfessionalSessionAuth": []
          }
        ],
        "summary": "Persist professional portal session",
        "tags": [
          "Professional Portal"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ]
}